+++ lib/algorand.transaction.acfg.ts
@@ -1,6 +1,6 @@
 import {AlgorandEncoder} from "./algorand.encoder.js"
 import {AssetParams} from "./algorand.asset.params.js";
-import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.js";
+import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.header.js";
 
 /**
  * This is used to create, configure and destroy an asset depending on which fields are set.
