+++ lib/algorand.transaction.afrz.ts
@@ -1,5 +1,5 @@
 import {AlgorandEncoder} from "./algorand.encoder.js"
-import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.js";
+import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.header.js";
 
 /**
  * Includes all fields in {@link TransactionHeader} and "type" is "axfer".
