+++ lib/algorand.transaction.pay.ts
@@ -1,5 +1,5 @@
 import {AlgorandEncoder} from "./algorand.encoder.js"
-import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.js";
+import {ITransactionHeaderBuilder, TransactionHeader} from "./algorand.transaction.header.js";
 
 /**
  * @category Transactions
