+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@algorandfoundation/algo-models",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
